dockerfiler: Easy Dockerfile Creation from R

Build a Dockerfile straight from your R session. 'dockerfiler' allows you to create step by step a Dockerfile, and provide convenient tools to wrap R code inside this Dockerfile.

See also

Author

Maintainer: Colin Fay contact@colinfay.me (ORCID)

Authors:

Vincent Guyader vincent@thinkr.fr (ORCID)

Josiah Parry josiah.parry@gmail.com (ORCID)

Sébastien Rochette sebastien@thinkr.fr (ORCID)