dockerfiler-package {dockerfiler}	R Documentation

dockerfiler: Easy Dockerfile Creation from R

Description

Build a Dockerfile straight from your R session. 'dockerfiler' allows you to create step by step a Dockerfile, and provide convenient tools to wrap R code inside this Dockerfile.

Author(s)

Maintainer: Colin Fay contact@colinfay.me (ORCID)

Authors:

Vincent Guyader vincent@thinkr.fr (ORCID)

Josiah Parry josiah.parry@gmail.com (ORCID)

Sébastien Rochette sebastien@thinkr.fr (ORCID)

See Also

Useful links:

https://github.com/ThinkR-open/dockerfiler

Report bugs at https://github.com/ThinkR-open/dockerfiler/issues

[Package dockerfiler version 0.2.2 Index]